Fix bugs and update unit tests
