+++ test/scenario_3/raider_example_3.yaml
@@ -11,7 +11,7 @@ aoi_group:
 height_group:
   height_levels: 0 100 500 1000
 los_group:
-  ray_trace: True  # Use projected slant delay by default
+  ray_trace: True  
   orbit_file: test/scenario_3/S1A_OPER_AUX_POEORB_OPOD_20181203T120749_V20181112T225942_20181114T005942.EOF
 runtime_group:
   output_projection:  4326
